added json file to read info and reformatted into diferent files
