+++ ipycmc/ipycmc/loadGeotiffs/variables.json
@@ -1,0 +1,13 @@
+{"required_start":["s3://", "S3://"],
+    "required_end":[".tif", ".TIF", ".tiff", ".TIFF"],
+    "defaults_tiler":{"tile_format":"png", "tile_scale":1, "pixel_selection":"first", "TileMatrixSetId":"WebMercatorQuad", "resampling_method":"nearest", "return_mask":"true", "rescale":"0%2C1000"},
+    "endpoints_tiler":{"maap-ops-workspace":"https://jqsd6bqdsf.execute-api.us-west-2.amazonaws.com", "maap-ops-dataset":"https://baxpil3vd6.execute-api.us-east-1.amazonaws.com"},
+    "tiler_extensions":{"single":"/cog/WMTSCapabilities.xml?","multiple":"/mosaicjson/WMTSCapabilities.xml?"},
+    "endpoint_published_data": "https://jqsd6bqdsf.execute-api.us-west-2.amazonaws.com",
+    "errors_tiler": {"not recognized as a supported file format.":"You have entered an invalid file path that does not exist or the TiTiler does not have access to.", "Access Denied":"The Tiler does not have access to the file path that you have provided, please give a maap s3 bucket or published data."},
+    "accepted_arguments_tiler":["minzoom", "maxzoom", "bidx", "expression", "nodata", "unscale", "color_formula", "colormap_name", "colormap"],
+    "mosaicjson_file":"mosaicjson.json",
+    "xml_beginning":"<",
+    "general_error_warning_tiler":"There was an error reading your link and Tiler gave the following error message: ",
+    "required_class_types_args_tiler":{"tile_scale":"int", "minzoom":"int", "maxzoom":"int", "bidx":"str", "expression":"str", "unscale":"bool", "rescale":"str", "color_formula":"str", "return_mask":"bool", "colormap":"str", "tile_format":"str", "TileMatrixSetId":"str", "resampling_method":"str", "colormap_name":"str", "pixel_selection":"str"}
+   }
